+++ examples/example.py
@@ -32,29 +32,32 @@ import runtask as rt           # task scheduler
 import time as tm              # time interface
 
 # set up the time scheduler
-scheduler = rt.RunTask(tick=0.1)
+sch = rt.RunTask(speed=1.0,tick=0.1)
 
-# a task printing run time and runs left, if not forever
-def task(scheduler):
+# a task printing run time and, if not forever, the runs left.
+def task(sch):
     now = tm.time()
-    task_id, runtime, runs = scheduler.task_info()
-    if runs == -1:
+    task_id, runtime, run_count, task_args = sch.task_info()
+    runs_left = sch.runs_left()
+    if runs_left == -1:
         print 'task','%d %12.3f %10.9f' % (task_id,runtime,now-runtime)
     else:
-        if runs != 1:
+        if runs_left == 1:
             print 'task','%d %12.3f %10.9f %d' \
-                % (task_id,runtime,now-runtime,runs)
+                % (task_id,runtime,now-runtime,runs_left), \
+                    'this is the last run'
         else:
             print 'task','%d %12.3f %10.9f %d' \
-                % (task_id,runtime,now-runtime,runs), 'this is the last run'
-            
+                % (task_id,runtime,now-runtime,runs_left)
+
+
 # task every 5 seconds, epoch aligned, forever
-scheduler.task(task,[scheduler],{},5.,0.0,-1)
+sch.task(task,[sch],{},sch.aligned(5.,0.0,-1))
 
 # two tasks every second, half second aligned,
 # the first forever, the second 5 times.
-scheduler.task(task,[scheduler],{},1.,0.5,-1)
-scheduler.task(task,[scheduler],{},1.,0.5,5)
+sch.task(task,[sch],{},sch.aligned(1.,0.5,-1))
+sch.task(task,[sch],{},sch.aligned(1.,0.5,5))
 
 # print a start message and start
 print 'Schedule 3 tasks for 10 seconds (system time) then terminate.'
@@ -63,10 +66,10 @@ print 'Task #1 is scheduled every second, aligned at half second, forever.'
 print 'Task #2 is scheduled every second, aligned at half second, for 5 times.'
 print 'id     runtime        sys-runtime run number'
 
-scheduler.start()
+sch.start()
 
 # wait 20 seconds then stop scheduler and exit.
 tm.sleep(10)
-scheduler.stop()
+sch.stop()
 
 #### END